Add calibrated, resumable live stacking and SNR analysis (#32)
* Add calibrated resumable live stacking

* Document live stacking and release builds

* Adopt Seiza 0.18.2 and harden calibration matching

* Adopt Seiza 0.18.3 calibration fixes

* Adopt Seiza 0.18.4 calibration fixes

* Test the release native runtime in CI

+++ src/Seiza.App/Models/ImageMetadata.cs
@@ -40,10 +40,10 @@ public sealed record ImageBackgroundDiagnostics(
     int RejectedResidual);
 
 public sealed record ImageStatistics(
-    int Minimum,
-    int Maximum,
+    double Minimum,
+    double Maximum,
     double Mean,
-    int Median,
+    double Median,
     double Mad);
 
 public sealed record ImageHistogram(
